refactor: move makeIdFromLabel to separate file
