+++ src/utils/index.ts
@@ -1,0 +1,2 @@
+export { ProductsSet } from './ProductsSet';
+export { makeIdFromLabel } from './MakeIdFromLabel';
